+++ src/main/java/apis/Manga/API/service/KontaktaufnahmeService.java
@@ -1,7 +1,7 @@
 package apis.Manga.API.service;
 
-import apis.Manga.API.Entety.Kontaktaufnahme;
 import apis.Manga.API.Repository.KontaktaufnahmeRepository;
+import apis.Manga.API.entity.Kontaktaufnahme;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -10,8 +10,11 @@ public class KontaktaufnahmeService {
 
     @Autowired
     private KontaktaufnahmeRepository kontaktaufnahmeRepository;
+    @Autowired
+    private AuthService authService;
 
     public Kontaktaufnahme getKontaktaufnahmeById(Long id) {
+        if (!authService.isAdmin()) return null;
         return kontaktaufnahmeRepository.findById(id).orElse(null);
     }
 
@@ -20,7 +23,9 @@ public class KontaktaufnahmeService {
     }
 
     public void deleteKontaktaufnahme(Long id) {
-        kontaktaufnahmeRepository.deleteById(id);
+        if (authService.isAdmin()) {
+            kontaktaufnahmeRepository.deleteById(id);
+        }
     }
 
     // Weitere Methoden hier hinzufügen, wenn benötigt.
