+++ src/main/java/apis/Manga/API/service/VideoService.java
@@ -1,7 +1,7 @@
 package apis.Manga.API.service;
 
-import apis.Manga.API.Entety.Video;
 import apis.Manga.API.Repository.VideoRepository;
+import apis.Manga.API.entity.Video;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -10,18 +10,22 @@ public class VideoService {
 
     @Autowired
     private VideoRepository videoRepository;
+    @Autowired
+    private AuthService authService;
 
     public Video getVideoById(Long id) {
+        if (!authService.isAdmin()) return null;
         return videoRepository.findById(id).orElse(null);
     }
 
     public Video createVideo(Video video) {
+        if (!authService.isAdmin()) return null;
         return videoRepository.save(video);
     }
 
     public void deleteVideo(Long id) {
-        videoRepository.deleteById(id);
+        if (authService.isAdmin()) {
+            videoRepository.deleteById(id);
+        }
     }
-
-    // Weitere Methoden hier hinzufügen, wenn benötigt.
 }
